Change `non_named::Semaphore::init_with` to not have `is_shared`.
The API is better this way, because now `anonymous::Semaphore::init_with` cannot panic, and now
generic uses of the trait method cannot possibly lead to a panic, and not having `is_shared` in
the trait method fits with OpenBSD not supporting it.

There is now the `unnamed::Semaphore::init_with` inherent method which still has `is_shared`,
because that's part of POSIX's unnamed semaphores and we want to expose that fully.  On OpenBSD,
`is_shared = true` will cause an error, and that's not our problem.